fix: refacto record component
