+++ js/__mocks__/react-native-device-info.js
@@ -1,0 +1,3 @@
+jest.mock('react-native-device-info', () =>
+	require('react-native-device-info/jest/react-native-device-info-mock'),
+)
