+++ js/package.json
@@ -63,7 +63,6 @@
 		"react-native-qrcode-scanner": "^1.6.0",
 		"react-native-qrcode-svg": "git+https://github.com/sakul-budhathoki/react-native-qrcode-svg.git#feature/circle-lines-qr-code",
 		"react-native-reanimated": "^1.13.2",
-		"react-native-responsive-screen": "^1.4.2",
 		"react-native-restart": "^0.0.20",
 		"react-native-safe-area-context": "^3.1.9",
 		"react-native-screens": "^2.17.1",
